+++ src/main/java/io/pakland/mdas/githubstats/domain/repository/RepositoryRepository.java
@@ -3,10 +3,6 @@ package io.pakland.mdas.githubstats.domain.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-/**
- * Add jdoc about the rep
- */
-
 @Repository
-public interface RepositoryRepository extends JpaRepository<io.pakland.mdas.githubstats.domain.Repository,Long> {
+public interface RepositoryRepository extends JpaRepository<io.pakland.mdas.githubstats.domain.Repository, Integer> {
 }
